pyproject.toml
[tool.poetry]
name = "fastapi-almbic-demo"
version = "0.1.0"
description = ""
authors = ["polymorphisma <shrawansunar.6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
opentelemetry-api = "^1.28.0"
opentelemetry-sdk = "^1.28.0"
opentelemetry-exporter-otlp = "^1.28.0"
opentelemetry-instrumentation-fastapi = "^0.49b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
